VM: Fix anonymous page-out after breaking on previous commit.

+++ kernel/include/kdk/nanokern.h
@@ -296,7 +296,7 @@ typedef struct kthread {
 	/* TCB */
 	uintptr_t tcb;
 	/* ID - only needed for userland threads. to move into ethread. */
-	uint16_t tid;
+	uintptr_t tid;
 } kthread_t;
 
 /*!
